+++ .gitignore
@@ -14,3 +14,5 @@ QubeSpec/detection/__pycache__/Detection.cpython-311.pyc
 QubeSpec/detection/__pycache__/__init__.cpython-311.pyc
 QubeSpec/detection/.DS_Store
 IFS_tutorial/.DS_Store
+QubeSpec/MSA/.DS_Store
+QubeSpec/detection/__pycache__/.DS_Store
