+++ setup.py
@@ -45,7 +45,7 @@ setup(name='QubeSpec',
       #      ],
       #},
       include_package_data=True,
-      package_data={'': ['detection/Detection.py','Model/Custom_model.py','Models/FeII_comp.py','Models/Halpha_models.py','Models/Halpha_OIII_models.py','Models/OIII_models.py','Models/QSO_models.py',
+      package_data={'': ['MSA/MSA.py', 'MSA/_init_.py','detection/Detection.py','Model/Custom_model.py','Models/FeII_comp.py','Models/Halpha_models.py','Models/Halpha_OIII_models.py','Models/OIII_models.py','Models/QSO_models.py',
           'Models/FeII_templates/FeII_Tsuzuki_opttemp.txt', 'Models/FeII_templates/bg92.con',
           'Models/FeII_templates/Veron-cetty_2004.fits', 'jadify_temp/red_table_template.csv',
           'jadify_temp/r100_jades_deep_hst_v3.1.1_template.yaml', 'jadify_temp/Temp_prism_clear_v3.0_extr3_1D.fits']},
